auth on index page

+++ pages/api/file.ts
@@ -39,8 +39,9 @@ export default (req: NextApiRequest, res: NextApiResponse) => {
 				'Custom Prompt': 'Say "You have not made a custom prompt in the editor yet!"'
 			};
 			const messageTypes: MessageType[] = (typeof fields.type === 'string') ? JSON.parse(fields.type) as MessageType[] : ['Linkedin invite'];
-			const auth = (typeof fields.storage === 'string') ? JSON.parse(fields.storage) : {};
-			if((auth.user !== 'admin') && (auth.pass !== process.env.ADMIN_PASS)) return res.status(400).json({ error: 'Invalid auth! User is not admin' });
+			let auth = (typeof fields.storage === 'string') ? JSON.parse(fields.storage) : {};
+			auth = JSON.parse(auth);
+			if((auth.user !== 'admin') || (auth.pass !== process.env.ADMIN_PASS)) return res.status(400).json({ error: 'Invalid auth! User is not admin' });
 			const workbook = xlsx.readFile(files.file.filepath);
 			const data = xlsx.utils.sheet_to_json(workbook.Sheets[workbook.SheetNames[0]]);
 			fs.unlinkSync(files.file.filepath);
@@ -86,6 +87,7 @@ const getResponses = async(data: any[], name: string, type: MessageType[], promp
 					useCache: 'false',
 					jsonmode: '%20'
 				  });
+				  console.log(personalizedInfo.data);
 		
 				const response = await openai.createCompletion({
 					model: "text-davinci-003",
